Fixing the intend issue
Also
1. Moving rsyslog tests to it's own file where it truly belongs
2. extracting the symlink dir as a separate method so it's easy to manage
